backmain/requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2020.6.20
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
cycler==0.10.0
distlib==0.3.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
email-validator==1.1.3
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.0.12
Flask==2.0.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
idna==3.2
importlib-metadata==1.7.0
itsdangerous==2.0.1
jieba3k==0.35.1
Jinja2==3.0.2
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mongoengine==0.23.1
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
openpyxl==3.0.9
pefile==2021.5.24
pika==1.2.0
Pillow==8.2.0
pipenv==2020.6.2
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pymongo==3.12.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
requests-file==1.5.1
selenium==3.141.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlparse==0.2.4
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.0.26
virtualenv-clone==0.5.4
Werkzeug==2.0.2
WTForms==2.3.3
zipp==3.1.0
